clean_sample_mean to utils
