+++ app/utils.py
@@ -1,9 +1,9 @@
 """
 Utility classes
 """
-import logging, os, signal, subprocess
+import csv, logging, os, signal, statistics, subprocess
 
-import cell, config
+import cell, config, exceptions
 
 logger = logging.getLogger('gage')
 
@@ -80,3 +80,44 @@ def remove_old_log_files():
     for path in old:
         logger.info(f'Removing log {path} as there are more than MAX_LOG_FILES ({config.MAX_LOG_FILES}).')
         os.remove(path)
+
+
+def writerow(row, data_csv_path):
+    """Write a row to the current csv file"""
+    if data_csv_path:
+        with open(data_csv_path, 'a') as f:
+            writer = csv.writer(f)
+            writer.writerow(row)
+    else:
+        logger.warning(f'DATA_CSV_PATH not avaliable, would have written: {row}')
+
+
+def clean_sample_mean(sample_func, ser, low, high, min_samples, max_attempts, max_std_dev):
+    """With a given sampling_func, sample and discard outliers"""
+    samples = [sample_func(ser)] # initial sample so that stdev doesn't yell at us for too few samples
+
+    for n in range(max_attempts):
+        samples.append(sample_func(ser))
+
+        cleaned_low_high = [s for s in samples if low <= s <= high]
+        try:
+            cleaned = [s for s 
+                        in cleaned_low_high 
+                        if abs(statistics.mean(cleaned_low_high) - s) < 2 * statistics.stdev(cleaned_low_high)]
+        except statistics.StatisticsError:
+            cleaned = []
+
+        if len(cleaned) >= min_samples:
+            if statistics.stdev(cleaned) < max_std_dev:
+                return statistics.mean(cleaned)
+    
+    if len(cleaned) < min_samples:
+        if low in samples:
+            raise exceptions.TooFewSamples(f'Too few cleaned samples that were not too low ({low})')
+        elif high in samples:
+            raise exceptions.TooFewSamples(f'Too few cleaned samples that were not too high ({high})')
+        else:
+            raise exceptions.TooFewSamples('Too few cleaned samples')
+    
+    stdev = round(statistics.stdev(cleaned), 2)
+    raise exceptions.SamplingError(f'Stdev ({stdev}) did not meet criteria ({max_std_dev}) in {max_attempts}')
